adding animations and no items view

+++ TodoList/Views/ListView.swift
@@ -12,23 +12,28 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+            
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
-        
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
             leading: EditButton(),
             trailing: NavigationLink("Add", destination:AddView()))
         
     }
-    
-    
 }
 
 struct ListView_Previews: PreviewProvider {
@@ -37,7 +42,6 @@ struct ListView_Previews: PreviewProvider {
             ListView()
         }
         .environmentObject(ListViewModel())
-    
     }
 }
